+++ src/lib/index.js
@@ -1,5 +1,6 @@
 import DescribedComponent from "./components/DescribedComponent";
 import DescribedButton from "./components/DescribedButton";
 import Glyphicon from "./components/Glyphicon";
+import AlertBox from "./components/AlertBox";
 
-export { DescribedComponent, DescribedButton, Glyphicon };
+export { DescribedComponent, DescribedButton, Glyphicon, AlertBox };